updated nabto cordova plugin version, adapted scripts to gracefully handle new ionic behavior

+++ scripts/ios-build.sh
@@ -11,13 +11,17 @@ fi
 
 ionic cordova platform add ios
 
-# work around for often seen problem when running emulator as of June 2017 (and it must be done in this odd way, even with previous installation using -g):
-# https://stackoverflow.com/questions/42350505/error-cannot-read-property-replace-of-undefined-when-building-ios-cordova
-(cd platforms/ios/cordova && npm install ios-sim@latest)
+if [ -e platforms/ios/cordova ]; then
 
-ionic cordova prepare ios
+  # work around for often seen problem when running emulator as of June 2017 (and it must be done in this odd way, even with previous installation using -g):
+  # https://stackoverflow.com/questions/42350505/error-cannot-read-property-replace-of-undefined-when-building-ios-cordova
+  (cd platforms/ios/cordova && npm install ios-sim@latest)
 
-# fix linker problem when using Nabto lib
-echo 'OTHER_LDFLAGS = -force_load $(BUILT_PRODUCTS_DIR)/libCordova.a -lstdc++' >> platforms/ios/cordova/build.xcconfig
+  ionic cordova prepare ios
 
-ionic cordova build ios
+  # fix linker problem when using Nabto lib
+  echo 'OTHER_LDFLAGS = -force_load $(BUILT_PRODUCTS_DIR)/libCordova.a -lstdc++' >> platforms/ios/cordova/build.xcconfig
+
+  ionic cordova build ios
+
+fi
